fixed some code for backend to work (in progress still)
