+++ package.json
@@ -14,7 +14,7 @@
     "dotenv": "^17.2.1",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.18.0",
-    "next": "^15.5.0",
+    "next": "^15.5.2",
     "next-pwa": "^5.6.0",
     "nodemon": "^3.1.10",
     "react": "^19.1.0",
