+++ src/app/api/auth/me/route.js
