+++ src/app/api/auth/onboarding/route.js
