+++ src/app/api/auth/register/route.js
